move authentication algorithms into security/
