+++ src/security/mod.rs
@@ -1,0 +1,6 @@
+mod des;
+pub use self::des::encrypt as des;
+
+mod md5;
+mod apple;
+pub use self::apple::apple_auth;
